refactor: move file to UI folder
